Update: Clean up the files related to directory commands
